requirements.txt
pip==23.3.1
dbt-core
dbt-snowflake~=1.7.0
mkdocs==1.5.3
mkdocs-awesome-pages-plugin==2.9.2
mkdocs-material==9.4.8
mkdocs-material-extensions==1.3
dbt-metricflow[snowflake]
